requirements.txt
PyJWT>=2.3.0
Flask-JSON>=0.3.4
bcrypt>=3.2.0
pyyaml
pyneurosdk2
Flask-SocketIO
opencv-python-headless 
Pillow
